+++ apps/researcher/src/app/globals.css
