+++ apps/researcher/src/app/not-found.tsx
@@ -1,14 +1,7 @@
-import {useLocale, useTranslations} from 'next-intl';
+import {useTranslations} from 'next-intl';
 
 export default function NotFound() {
-  const locale = useLocale();
   const t = useTranslations('NotFound');
 
-  return (
-    <html lang={locale}>
-      <body>
-        <h1>{t('title')}</h1>
-      </body>
-    </html>
-  );
+  return <h1>{t('title')}</h1>;
 }
